JS optimisations for Std.is.
Generate exact code for typed cases of Std.is(x, T) when T is
known at compile time to be String/Int/Bool/Dynamic/Float.

This is handled as an optimization step specifically for JS
which transforms calls to js.Boot.__instanceof appropriately
before DCE kicks in, ensuring that if the only isntances of Std.is
is to one of these known types, then js.Boot remains excluded from
the build nicely.