Compiler.hx keep macro fix for case (some more)
With :
src/Game.hx
src/game/

And
keep(“Game”)

The following error is thrown :
--macro:1: character 0 : Invalid package : Game should be game

This is because FileSystem is not case sensitive for .exists(), it
would find the package folder for Game under game/ and suggest
isValidDirectory is true. On platforms with insensitive paths, this is
an error.

This commit checks the case against the parent folder of the searched
keep in order to ensure case is respected before it tries to set
isValidDirectory.