Throw an error if eelse isn't defined.
The Haxe compiler optimizes "`if(true) 1 else 0`" to 1, which can then be used as a value. However, "`if(true) 1`" results in `Void`, rather than 1, meaning it cannot be used as a value.

To match this behavior, `getValue()` should only return a value for an `EIf` expression if `eelse` is defined, regardless of what `econd` evaluates to.