Save big files using Http
Adds the posibility to set an haxe.io.Output to save the response into a file for example. So if you want to download a big file, you don't have much memory consumption.
Also adds the posibility to cancel the ouput reading from socket, allowing to cancel the request (if run it in a thread for example).