Lua: invoking instance methods correctly
Well, this is the second time I've added in some commits to a
misc/cleanup that ended up needing a small note on their own.

Lua has a method for passing a "self" (i.e. this) parameter to certain
methods, but it requires a different dot syntax to invoke it... instead
of "foo.bar" on a given instance, you need to use "foo:bar".  This means
we need to catch function calls and inspect if they're instance methods
or not (static methods need to stay unchanged).

I tried to fix this before, but this CL updates the behavior here a bit better.