Lua : Fix weird situation with switches on enums.
It looks like Haxe generates conditionals in a strange way with certain
enums... it seems like it will generate part of the way as an if
statement, while the rest of the expression will be the default part of
a Switch statement.  That's a bit hard to explain, even with haxe code.

So, hopefully the code change below gives some idea... Both if and
switch statements generate lua if statements.  In the switch statement,
if I can detect that only the default statement exists (no extra
switches), then I avoid creating the complementary "else" statement
there.