Lua : refactor OOP
Previously I was using a hacked together oop implementation for lua I
came up with.   It worked ok for methods, but didn't quite get the
fields right in certain cases, specifically with certain usages of
super().

So, I've gone back to the drawing board with something more robust. The
biggest change is that in the generated Lua code, all constructors now
require at least a single argument, typically an anonymous table.

```lua
Foo.new({});
```
Haxe lua users won't have to worry about this, the constructor is
generated automatically with the base instance specified.

The main benefit of this is that rather than constructing a new instance
internally, this anonymous table serves as the base object from which to
extend the fields and methods.  This allows a very straightforward
implementation for super(), which can be a simple alias for the extended
class constructor, called on the current instance.  It will emulate the Haxe
specification for super perfectly.  This will also be nice for
extending base lua objects with haxe-specific functionality (such as
arrays).

I was hoping to put this off a bit longer, but it's tripping up some of
the tests.