Lua : Fix for null values in string concatenation
Briefly, Lua doesn't allow for arbitrary string concatenation with nil
values (equivalent to haxe null).  Haxe expects this (or at least, it
tests for it in the unit tests).  In order to support this, we have to
check for an add operation where at least one value is a string, and
then wrap both arguments in a method that will guarantee a string
output.  For the Lua target, this method should be Std.string.  It
is generated accordingly in this case using the logic here.