Lua: fix up a lua specific $bind method.
In haxe, it's possible to separate an instance method from the instance
that calls it.  In that case, it's necessary to encapsulate the function
in a closure, and retain the original instance inside of it.

The lua version can look pretty similar to the js version, but it needs
a few modifications.  Also, the original js implementation was
completely inlined/optimized.  I wanted to break it apart and think
about it first, so I stashed it on lua.Boot.bind for now.  It's not 100%
optimal, but it appears to work for now.