Lua : Hoist __static__ variable declarations
In Haxe, a __static__ method on a class has special behavior.  It
functions as a low level initialization for the class.  It's not
commonly used, but it's a necessary part of the language for many
purposes.

One of these purposes is the __static__ method on the Boot class which
initializes the base abstracts for Int, Bool, etc.  This works fine in
js, which can automatically hoist any variable declarations.  However,
in lua the local instances that carry type information are not
avaialable in any of the code prior to the __static__ declaration. They
won't work.

It's also not a good idea to move the __static__ declarations to the top
of the file... there is a lot of other implementations for String and
Array missing, so any code there would have crippling limitations.

So, the workaround here is to hoist *just* the local variable declarations in
__static__ blocks to the top.  Then, the Array/String/Class definitions
are provided, then the actual __static__ implementation, then the
individual static field generation, and finally the static main
invocation.