Lua : Use 0-based indexing for enum values.
Haxe enums are typically array run-time values, and the expression is
generated further upstream in the ast generator.  I'm not sure if it's a
good idea to try and make that use 1-based indexing for raw lua tables,
or if we should just use 0-based indexing here.  For now, I'm using
0-based indexing, and adding a todo for later.