Add filename & line # to HX_STACK_FRAME's fullName
hxcpp's `HX_STACK_FRAME` takes a `const char* fullName` that are used in Debug.h's stack class. It looks like the only thing using these (in hxcpp) is the `Profiler` and `Telemetry` classes.

In the case of lambdas, the class seems to always be `*` while the function name is of the form `_Function_#_#`, creating a fullName, e.g. `*._Function_1_2`

This is unfriendly for debugging, locating these lambdas is difficult.

This commit adds a `(filename:line#)` suffix to the fullName when class is `*` (this should only be lambdas), so that the fullName becomes e.g. `*._Function_1_2 (openfl/_internal/text/TextEngine.hx:634)`

Again, this is used in HX_STACK_FRAME calls, e.g.:

```
			HX_STACK_FRAME("*","_Function_1_2",0x5200ed38,"*._Function_1_2 (openfl/_internal/text/TextEngine.hx:634)","openfl/_internal/text/TextEngine.hx",634,0xdff93b7f)
```

This results in easier profiling from Profiler or Telemetry.

Possible modifications to this PR:
- The syntax of the addition is not important, as long as it conveys filename and line number.
- Absolute path name isn't necessarily important (could just be `TextEngine.hx`, but full path was already available).

@hughsando - thoughts? Does this cause problems for any tools you know of? Does anything use stack.fullName outside Profiler and Telemetry?