[cs] Fix Date.fromTime (see #4584)
This fixes failing test added in eca09499dc7b101eaacd2d1ad61f60ddde506e43 on my machine.
I expected it to fail on Travis as well, but it didn't for some reason. Probably it has something to do
with sytem clock set to UTC on *nixes, let's see if that commit will fail on AppVeyor (and fixed by this commit).