Convert underscores to hyphens in php client headers
All incoming headers are transformed by php to an underscore format in $_SERVER. For example `Content-type` is found in `$_SERVER['CONTENT_TYPE']`. To correctly get the header keys, the underscores need to be converted to hyphens again. Headers containing underscores are very rare and ignored in most webservers (apache, nginx).