Fix haxe.Template resolve method (resolves #5287)
The `resolve` method looks for matches in the template context, and a context stack before finally checking globals.

If a match did not resolve on `template.context`, the code wrongfully overwrote the name of the string key you were looking for. This would eventually lead to a wrongful `Reflect.hasField(ctx,null)` call that crashes Neko. You cannot pass a null string value to `Reflect.hasField` on Neko

This fix looks like the correct behavior