This is a work in progress change that enables multiple returns in Lua
Lua supports multiple returns and uses them heavily in its standard
library. Currently, Haxe does not support multiple returns, so
accessing the return values from these methods is problematic, and
describing the result of these methods in the haxe type system is
impossible.

```lua
local begin,end = string.find
```

Currently, the multi-return methods are treated as only returning one
result, and typing is provided for that return only.

I've tried different approaches at remedying the situation.  The first
attempt simply wraps the results of multiple return methods in a table,
enabling the values to be passed as a singe variable reference.
This makes it easier to access subsequent return values, but it carries
significant overhead.

The second approach was to create a special abstract typedef metadata,
that would  mark the type as containing a multiple return value:

```haxe
@:multiReturn typedef StringFind = {
    begin : Int,
    end : Int
}
```

This enabled the type to be "unpacked" into a sequence of variable
declarations:

```haxe
var strfind = lua.NativeStringTools.find("foo bar", "foo");
trace(strfind.begin);
```

```lua
local _hx_1, _hx_2 = string.find("foo bar", "foo");
haxe.Log.trace(_hx_1);
```

When a variable is declared to accept the result of a multireturn
method, it will populate a new "v_multi" field with a list of temporary
variable names corresponding to each of its fields.  Each time those fields
are requested, they are resolved against the v_multi index, and that
temporary variable name is used instead of the field expression.

While the technique worked in principle, a problem happened with the
structural typedefs unifying too readily with other typedefs.  This
will likely cause a lot of confusion when working with multi return
typedefs in conjunction with other normal typedefs.

This next approach attaches the @:multiReturn metadata to a special
class instance instead, but it works more or less the same, and does not
have nearly the same problems with unification since it uses the more
structured inheritance semantics.

The end result is the ability to provide quite robust multi return
semantics without having to wrap the values.

```haxe

class Main {
    static function test(){
        trace(find("foo bar", "foo").begin + " is the value for k"); // we can access the fields of the return immediately
        var k = find("foo bar", "foo"); // we can assign to a new variable, which has special semantics.
        var l = k.begin; // we can assign out of the multireturn type to a new variable
        trace(l + " is the value for l");
        trace(k.end + " is the value for k.end");

        // the multireturn reference is just a list of variables,
        // so it doesn't print correctly... yet.
        trace(k + " is the value for k");

        return k; // but, we can return multireturn types
    }
    static function main() {
        var f = test();
        // get the multi return value from test, the semantics are preserved.
        trace(f.end + " is the value for f.end");
    }
    static function find(str1:String, str2:String) : StringFind {
        return untyped lua.NativeStringTools.find(str1, str2);
    }

}

// This is our special multireturn wrapper value
// the order of the declared instance variables is the order
// of the multiple returns from the corresponding method.
@:multiReturn extern class StringFind {
    var begin : Int;
    var end : Int;
}

```