[lua] Add multireturn behavior handling for lua externs
This is the third version of multireturn behavior.  This version detects
when a variable is assigned from the result of a Meta.Multireturn method
call.  In this case, the Meta.MultiReturn meta is added to the v_meta
field of the local variable, along with a local mapping of the field
names to the temp name aliases.

This method is in contrast to the previous attempt, which used a special
v_multi field on the variable itself.