Fix HaxeFoundation/haxe#5941.
The zlib_issue was caused because the OCaml garbage collector moved
the zlib data allocated on the heap while zlib kept backreference. This
commit uses OCaml's custom memory allocation function to avoid this
issue. This also adds a .editorconfig file for *.c and a few comments
in extc_stubs.c.