AS3 target Compile error with missing iterators. (#6187)
* Update WeakMap.hx

Error when targeting as3. Generates NativePropertyIterator.as and NativeValueIterator.as but __foreach__ __forin__ missing.  Better to not generate at all, since those impl are not used.

* Update ObjectMap.hx

Error when targeting as3. Generates NativePropertyIterator.as and NativeValueIterator.as but __foreach__ __forin__ missing.  Better to not generate at all, since those impl are not used.