Add support for toString metadata.
toString metadata existed but was not used, now it can be used to
annotate methods to provide DCE information about which types
will be converted to string and thus, its `toString` method should be
preserved.

This metadata mechanism replaces the previous way of adding this
information directly into the DCE code.

Fixes #6841