[display] fix display offset inconsistency
It turns out that modes like definition, references and hover really work differently from completion: A good example of this is `foo.|field` with the cursor being at the `|`. For completion, you would want to get the fields of `foo`, whereas with definition you would want to find the definition of `field`.

closes #6923