[std] remove @:require(haxe_ver) metas
This was intended for API doc generation, but these metas were never added to new APIs consistently and dox was also never updated to show this info. They serve no purpose otherwise, since the standard lib is not backwards-compatible with older Haxe versions.

It makes sense to remove them now because the checks start failing if haxe_ver is changed to a SemVer (see #8302).

Starting with Haxe 4.1, it might be nice to start using JavaDoc @since tags (respected by both dox and vshaxe).