[cs][java][jvm] clean up -main handling a little (module-statics preparation) (#9230)
* [cs][java][jvm] clean up -main handling a little (module-statics preparation)

these are the only targets that actually rely on the com.main_class and treat it as an actual class path. to support module-level -main i need to treat that as module path, this change will make it easy to change the real main class because i will only need to change the get_entry_point implementation.

also this fixes genjvm generating entry-point main methods for any "main" method :-)

* [jvm] fix entry-point logic a bit

only matters when main class will be something different (e.g. module-statics), but still