Use underscore defines internally
Move all define conversions to interfaces, and internally a flag is
represented using the underscore version only.

This prevents duplicate defines appearing in a bunch of places, and now
there are no longer tiny differences in the effects of `-D flag-a` and
`-D flag_a`. Same is true for macro functions for accessing defines.