Avoid calling external processes through the system shell
This change allows Haxe to be used even if the system shell is unavailable.

Certain ocaml functions which call external processes do so through the system
shell. On systems where the shell is unavailable, this makes Haxe unusable.

These problematic functions are `Unix.system`, the `Unix.open_process_*` family
of functions, and `Sys.command`.

Places where the shell was previously accessed:

1. When using the `-lib` argument to the compiler.
- `Haxe.add_libs` (calls `haxelib path {libs}`)

2. When using the `-cmd` argument to the compiler.
- `Haxe.run_command` (command is normally passed to the shell)

3. When generating code for certain targets.
- `Gencpp.write_build_options` (calls `haxelib path hxcpp`)
- `Gencpp.generate_source` (calls `haxelib run hxcpp ...`)
- `Gencs.generate` (calls `haxelib run hxcs ...`)
- `Genhl.generate` (calls `haxelib run hashlink ...`)
- `Genjava.generate` (calls `haxelib run hxjava ...`)
- `Genneko.generate` (calls `nekoc ...`)

4. By hl interpreter and eval, to check the system name.
- `Hlinterp.load_native`, for `sys_string` (calls `uname` on unix)
- `EvalStdLib.StdSys` for system name (calls `uname` on unix)

Out of all of these, most don't need to be interpreted by the shell. All the
calls to `haxelib`, `nekoc`, and `uname` are known in advance, and those
processes can be started directly instead. The only one that actually needs to
be interpreted by the shell is (2), using `-cmd`.

As a replacement for the `Unix.open_process_*` family of functions, the Unix
module provides `Unix.open_process_args_*` as well. This set of functions does
not pass through the shell, but uses fork/exec/posix_spawn on unix and
CreateProcess on Windows. Arguments are passed as an array of strings, and
given that the shell is not involved, the strings should not be escaped like
they would be if passed to the shell. The first argument in the array should
be the name of the program.

For the purpose of finding the full path, a simple path-searching function has
been added in the new process_helper.ml. This path searcher does not function
exactly as the built-in path-searching capabilities of system shells.

Unfortunately, `Unix.open_process_args_*` is only available beginning with
ocaml 4.08, so a backport that wraps around `Unix.create_process_env` has also
been added in process_helper.ml.