Cleanup minor environment variable issues (#10464)
* [python] Ensure putEnv properly removes variables

Requires os.unsetenv extern

* [tests] Add extra testing for removing env variables

Previously the tests did not ensure that the variables would
be updated in child processes

* [python] Fix Sys.putEnv for python 3.8 and below

Add a try catch block so that this function doesn't throw when targeting
python versions prior to 3.9, as before that `os.unsetenv` is not
guaranteed to exist

* Document putEnv(var, null) limitation on python

* [python] Remove internal environment variable map

Fixes #10469

Sys.putEnv() now modifies python's `os.environ` dictionary rather than
using `os.putenv` or `os.unsetenv` and modifying an internal map. This
is a better choice as modifying the dictionary calls these functions
automatically if they exist.
This also means that changes made via the Haxe api will be visible also
via python's `os.environ`.

As a result, Sys no longer needs to store an internal map of environment
variables. As of #10460, a copy is made of the map
anyway everytime `Sys.environment()` is called. This means we might as
well directly loop through python's `os.environ` which takes into
account modifications made via the python
api, rather than looping through an internal copy which does not, and
takes the same amout of time.

* [cs] Remove internal copy of environment variables

Fixes #10469

Now, modifications to the environment directly via the
native C# api will be visible in 'Sys.environment()`

* [tests] Update sys tests accordingly

* [python] Allow removing non-existent variable

* [tests] Test removal of non-existent variable

* [python] Remove warning about putEnv(name, null)

This should not be an issue now that Sys.putEnv modifies `os.environ`
directly

* [tests] Only test uppercase environment variables

For compatibility with python

* [tests] Test getEnv case-insensitivity on windows

* [php] Remove incorrect environment variables

Fixes #10471

On windows, ensure only correct and up to date environment variables are
present in `Sys.environment()`

* [sys] Document Sys.environment() capitalization