bring back support for absolute paths in find_file
This is a bit messy now because we want to track which class path we used to resolve a file, but in the case of absolute paths we didn't really use a class path. It makes sense to introduce an absolute class path like here, but we already use the empty string as the cwd class path too.