Run toString(null) for explicit nullable abstract
If the abstract is declared over Null<T>, then null is part of the type
so the user should be able to handle the null case within the toString
method. It is also useful to be able to have custom empty values for
an abstract that wraps a nullable type.

This is also in-line what null safety has to say, as it forces the user
to add a null check to toString when the concrete type is Null<T>.