[hlc] Move hlc_main.c #include below source files
hlc_main.c includes windows.h, which defines macros that can conflict
with user code. By placing it below the source file includes, we avoid
this being an issue.

Note, this has to be combined with the hashlink patch to remove
windows.h from hl.h to have any benefit.